Added method appendSearchableComboBox
Similar to the other various combox appending methods, except this version additionally installs a JIDE Searchable component onto it, allowing the user to type in a partial value to help locate a particular entry they may be looking for.  Helpful when the list of values may be long and/or out of order.